Fix could not convert template argument error
When compiling to a x86-64 target using gcc, clang and msvc the following error appears:

In file included from spark/include/SPARK_Core.h:38,
                 from spark/src/Core/IO/SPK_IO_Buffer.cpp:22:
spark/include/Core/SPK_DescriptionDefines.h:266:123: error: could not convert template argument ‘&SPK::Group::getCapacity’ from ‘size_t (SPK::Group::*)() const’ {aka ‘long unsigned int (SPK::Group::*)() const’} to ‘unsigned int (SPK::Group::*)() const’
  266 |                                                                 ::template store<SPK_UNPACK_GETTERS_(_spk_obj,__VA_ARGS__)>,\
      |
spark/include/Core/SPK_DescriptionDefines.h:154:25: note: in definition of macro ‘_spk_description_body’
  154 |                         __VA_ARGS__ \
      |                         ^~~~~~~~~~~
spark/include/Core/SPK_Group.h:373:25: note: in expansion of macro ‘spk_attribute’
  373 |                         spk_attribute(unsigned int, capacity, reallocate, getCapacity);
      |                         ^~~~~~~~~~~~~

On a x86-64 target (using those compilers) the size of size_t is 64bit and the size of
unsigned int is 32bit hence the 'could not convert' error.

The ParticleData struct field nbParticles was also set to unsigned int fix the packing.

+++ include/Core/SPK_Group.h
@@ -91,13 +91,13 @@ namespace SPK
 
 		bool isEnabled(Param param) const;
 
-		size_t getNbParticles() const;
-		size_t getCapacity() const;
+		unsigned int getNbParticles() const;
+		unsigned int getCapacity() const;
 
 		Particle getParticle(size_t index);
 		const Particle getParticle(size_t index) const;
 
-		void reallocate(size_t capacity);
+		void reallocate(unsigned int capacity);
 		void empty();
 
 		void addEmitter(const Ref<Emitter>& emitter);
@@ -401,8 +401,8 @@ namespace SPK
 		{
 			bool initialized;
 
-			size_t nbParticles;
-			size_t maxParticles;
+			unsigned int nbParticles;
+			unsigned int maxParticles;
 
 			// Particles attributes
 			Vector3D* positions;
@@ -734,12 +734,12 @@ namespace SPK
 		return particleData.parameters[param] != NULL;
 	}
 
-	inline size_t Group::getNbParticles() const
+	inline unsigned int Group::getNbParticles() const
 	{
 		return particleData.nbParticles;
 	}
 
-	inline size_t Group::getCapacity() const
+	inline unsigned int Group::getCapacity() const
 	{
 		return particleData.maxParticles;
 	}
